Fixed timeout error
Fixed an oversight where if Discord replied with "503 service unavailable" (or any other error) during "interaction.showModal(modal)" when a user tries to answer a captcha, the bot would hard crash. Bot no longer crashes & instead retries the "interaction.showModal(modal)" action until it completes, or, as a fallback, it shows a manual "Retry" button to the user.